Moved pages to Pages folder, Implemented POST data function as a Redux saga side effect
