+++ src/state/sagas/saga.js
@@ -1,6 +1,6 @@
-import { call, takeLatest,put } from "redux-saga/effects";
-import {getCards,setCards} from "../slices/cardSlice";
-import {getDocs,collection } from "firebase/firestore";
+import { call, takeEvery,put } from "redux-saga/effects";
+import {getCards,setCards,postCard} from "../slices/cardSlice";
+import {getDocs,collection,addDoc } from "firebase/firestore";
 import {db} from "../../db/firebase";
 
 async function requestGetCards(){
@@ -12,16 +12,31 @@ async function requestGetCards(){
     return cards;
 }
 
+async function postDiaryCard({name,title,description}){
+    await addDoc(collection(db,"diarycards"),{
+       name,title,description
+    });
+}
+
+function* handlePostDiaryCard({payload}){
+    try{
+        yield call(postDiaryCard,payload);
+        console.log("posted the document")
+    }catch(err){
+        console.log(err);
+    }
+}
+
 function* handleGetCards(){
     try{
-        const res = yield call(requestGetCards)
+        const res = yield call(requestGetCards);
         yield put(setCards({res}));
-       
     }catch(err){
         console.log(err);
     }
 }
 
 export function* rootSaga(){
-    yield takeLatest(getCards.type,handleGetCards);
+    yield takeEvery(getCards.type,handleGetCards);
+    yield takeEvery(postCard.type,handlePostDiaryCard);
 }
